Revert new comparison logic
